# ⚡️ Speed up function `under_non_alpha_ratio` by 76%