--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ### Enhancements
 
+- Speed up function `under_non_alpha_ratio` by 76% (codeflash)
+
 ### Features
 
 ### Fixes
 
 - **change short text language detection log to debug** reduce warning level log spamming
-- Speed up function `under_non_alpha_ratio` by 76% (codeflash)
+
 
 ## 0.18.13